+++ client/v2/autocli/flag/builder.go
@@ -20,8 +20,6 @@ import (
 	"cosmossdk.io/client/v2/internal/flags"
 	"cosmossdk.io/client/v2/internal/util"
 	"cosmossdk.io/core/address"
-
-	"github.com/cosmos/cosmos-sdk/runtime"
 )
 
 const (
@@ -55,8 +53,8 @@ type Builder struct {
 
 	// Address Codecs are the address codecs to use for client/v2.
 	AddressCodec          address.Codec
-	ValidatorAddressCodec runtime.ValidatorAddressCodec
-	ConsensusAddressCodec runtime.ConsensusAddressCodec
+	ValidatorAddressCodec address.ValidatorAddressCodec
+	ConsensusAddressCodec address.ConsensusAddressCodec
 }
 
 func (b *Builder) init() {