+++ x/staking/simulation/proposals_test.go
@@ -11,6 +11,7 @@ import (
 	"cosmossdk.io/x/staking/simulation"
 	"cosmossdk.io/x/staking/types"
 
+	codectestutil "github.com/cosmos/cosmos-sdk/codec/testutil"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/address"
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
@@ -23,7 +24,7 @@ func TestProposalMsgs(t *testing.T) {
 
 	ctx := sdk.NewContext(nil, true, nil)
 	accounts := simtypes.RandomAccounts(r, 3)
-
+	addressCodec := codectestutil.CodecOptions{}.GetAddressCodec()
 	// execute ProposalMsgs function
 	weightedProposalMsgs := simulation.ProposalMsgs()
 	assert.Assert(t, len(weightedProposalMsgs) == 1)
@@ -34,11 +35,15 @@ func TestProposalMsgs(t *testing.T) {
 	assert.Equal(t, simulation.OpWeightMsgUpdateParams, w0.AppParamsKey())
 	assert.Equal(t, simulation.DefaultWeightMsgUpdateParams, w0.DefaultWeight())
 
-	msg := w0.MsgSimulatorFn()(r, ctx, accounts)
+	msg, err := w0.MsgSimulatorFn()(r, ctx, accounts, addressCodec)
+	assert.NilError(t, err)
 	msgUpdateParams, ok := msg.(*types.MsgUpdateParams)
 	assert.Assert(t, ok)
 
-	assert.Equal(t, sdk.AccAddress(address.Module("gov")).String(), msgUpdateParams.Authority)
+	addr, err := addressCodec.BytesToString(address.Module("gov"))
+	assert.NilError(t, err)
+
+	assert.Equal(t, addr, msgUpdateParams.Authority)
 	assert.Equal(t, "GqiQWIXnku", msgUpdateParams.Params.BondDenom)
 	assert.Equal(t, uint32(213), msgUpdateParams.Params.MaxEntries)
 	assert.Equal(t, uint32(300), msgUpdateParams.Params.HistoricalEntries)